вынес данные комментов в отдельный файл и добавил сворачивание/разворачивание этих комментов
