+++ components/SideComments/CommentItem.tsx
@@ -1,0 +1,42 @@
+import React from 'react'
+import styles from './SideComments.module.scss'
+import Link from 'next/link'
+import Image from 'next/image'
+
+interface CommentItemProps {
+  user: {
+    id: number
+    fullname: string
+  }
+  text: string
+  post: {
+    id: number
+    title: string
+  }
+}
+
+export const CommentItem: React.FC<CommentItemProps> = ({ user, text, post }) => {
+  return (
+    <div className={styles.commentItem}>
+      <div className={styles.userInfo}>
+        <Image
+          src="https://leonardo.osnova.io/598fc957-a3f6-598c-b6f9-a033c3941d12/-/scale_crop/64x64/-/format/webp/"
+          alt="User avatar"
+          height="32px"
+          width="32px"
+        />
+        <Link href={`/profile/${user.id}`}>
+          <a>
+            <b>{user.fullname}</b>
+          </a>
+        </Link>
+      </div>
+      <p className={styles.text}>{text}</p>
+      <Link href={`/news/${user.id}`}>
+        <a>
+          <span className={styles.postTitle}>{post.title}</span>
+        </a>
+      </Link>
+    </div>
+  )
+}
