+++ components/SideComments/SideComments.module.scss
@@ -1,6 +1,9 @@
 .root {
   width: 320px;
   margin-bottom: 2px;
+  position: sticky;
+  top: 90px;
+  margin-left: 40px;
 
   h3 {
     display: flex;
@@ -15,6 +18,16 @@
       height: 22px;
       margin-left: 8px;
     }
+  }
+}
+
+.rotated {
+  width: 100px;
+  transform: rotate(-90deg) translateX(-50px);
+  transform-origin: right;
+
+  svg {
+    transform: rotate(270deg);
   }
 }
