+++ components/SideComments/index.tsx
@@ -1,75 +1,24 @@
 import React from 'react'
 import ArrowRightIcon from '@mui/icons-material/NavigateNextOutlined'
-import Image from 'next/image'
 
+import data from '../../data'
 import styles from './SideComments.module.scss'
-
-const items = [
-  {
-    user: {
-      fullname: 'Вася Пупкин',
-    },
-    text: 'Теперь, каждое рабочее утро, после кровати, я перекладываюсь туда спать ещё на часок. Ну и…',
-    post: {
-      title: 'Какая у вас дома ванна?',
-    },
-  },
-  {
-    user: {
-      fullname: 'Вася Пупкин',
-    },
-    text: 'Теперь, каждое рабочее утро, после кровати, я перекладываюсь туда спать ещё на часок. Ну и…',
-    post: {
-      title: 'Какая у вас дома ванна?',
-    },
-  },
-  {
-    user: {
-      fullname: 'Вася Пупкин',
-    },
-    text: 'Теперь, каждое рабочее утро, после кровати, я перекладываюсь туда спать ещё на часок. Ну и…',
-    post: {
-      title: 'Какая у вас дома ванна?',
-    },
-  },
-]
-
-interface CommentItemProps {
-  user: {
-    fullname: string
-  }
-  text: string
-  post: {
-    title: string
-  }
-}
-
-const CommentItem: React.FC<CommentItemProps> = ({ user, text, post }) => {
-  return (
-    <div className={styles.commentItem}>
-      <div className={styles.userInfo}>
-        <Image
-          src="https://leonardo.osnova.io/598fc957-a3f6-598c-b6f9-a033c3941d12/-/scale_crop/64x64/-/format/webp/"
-          alt={'#'}
-        />
-        <a href="#">
-          <b>{user.fullname}</b>
-        </a>
-      </div>
-      <p className={styles.text}>{text}</p>
-      <a href="#">
-        <span className={styles.postTitle}>{post.title}</span>
-      </a>
-    </div>
-  )
-}
+import { CommentItem } from './CommentItem'
+import clsx from 'clsx'
 
 export const SideComments = () => {
+  const [visible, setVisible] = React.useState(true)
+
+  const toggleVisible = () => {
+    setVisible(!visible)
+  }
+
   return (
-    <div className={styles.root}>
-      <h3>
+    <div className={clsx(styles.root, !visible && styles.rotated)}>
+      <h3 onClick={toggleVisible}>
         Комментарии <ArrowRightIcon />
       </h3>
+      {visible && data.comments.popular.map((obj) => <CommentItem key={obj.id} {...obj} />)}
     </div>
   )
 }
